add /file route

+++ server.js
@@ -5,21 +5,28 @@ const port = 4000;
 
 http.createServer((req, res) => { 
   console.log("request!!")
-  const filePath = 'file.log';
-
-  fs.unlinkSync(filePath);
+  
+  const filePath = 'file.log';  
 
   res.writeHead(200, {'Content-Type': 'text/plain'});
 
-  req.on('data', chunk => {
-      console.log(`Data chunk available: ${chunk}`);
+  if (req.url === '/file') {
+    const data = fs.readFileSync(filePath, 'utf8');
+    res.end(data);
+  }        
 
-      fs.appendFile(filePath, chunk, err => {
-        if (err) {
-          console.error(err);
-        }
-      });
-      res.end();
+  req.on('data', chunk => {
+    if (fs.existsSync(filePath)) fs.unlinkSync(filePath);
+    
+    console.log(`Data chunk available: ${chunk}`);
+
+    fs.appendFile(filePath, chunk, err => {
+      if (err) {
+        console.error(err);
+      }
+    });
+
+    res.end();
   });
     
 }).listen(port);
